Modularize back-end logic
